+++ kogito-jbpm/jbpm-persistence-jpa/src/test/java/org/jbpm/persistence/session/SingleSessionCommandServiceTest.java
@@ -112,7 +112,7 @@ public class SingleSessionCommandServiceTest extends AbstractBaseTest {
         SingleSessionCommandService service = new SingleSessionCommandService( kbase,
                                                                                config,
                                                                                env );
-        int sessionId = service.getSessionId();
+        Long sessionId = service.getSessionId();
 
         StartProcessCommand startProcessCommand = new StartProcessCommand();
         startProcessCommand.setProcessId( "org.drools.test.TestProcess" );
@@ -226,7 +226,7 @@ public class SingleSessionCommandServiceTest extends AbstractBaseTest {
         SingleSessionCommandService service = new SingleSessionCommandService( kbase,
                                                                                config,
                                                                                env );
-        int sessionId = service.getSessionId();
+        Long sessionId = service.getSessionId();
 
         UserTransaction ut = (UserTransaction) new InitialContext().lookup( "java:comp/UserTransaction" );
         ut.begin();
@@ -428,7 +428,7 @@ public class SingleSessionCommandServiceTest extends AbstractBaseTest {
         SingleSessionCommandService service = new SingleSessionCommandService( ruleBase,
                                                                                config,
                                                                                env );
-        int sessionId = service.getSessionId();
+        Long sessionId = service.getSessionId();
         StartProcessCommand startProcessCommand = new StartProcessCommand();
         startProcessCommand.setProcessId( "org.drools.test.TestProcess" );
         RuleFlowProcessInstance processInstance = (RuleFlowProcessInstance) service.execute( startProcessCommand );
@@ -615,7 +615,7 @@ public class SingleSessionCommandServiceTest extends AbstractBaseTest {
         SingleSessionCommandService service = new SingleSessionCommandService( kbase,
                                                                                config,
                                                                                env );
-        int sessionId = service.getSessionId();
+        Long sessionId = service.getSessionId();
         StartProcessCommand startProcessCommand = new StartProcessCommand();
         startProcessCommand.setProcessId( "org.drools.test.TestProcess" );
         ProcessInstance processInstance = service.execute( startProcessCommand );
@@ -720,7 +720,7 @@ public class SingleSessionCommandServiceTest extends AbstractBaseTest {
         SingleSessionCommandService service = new SingleSessionCommandService( kbase,
                                                                                config,
                                                                                env );
-        int sessionId = service.getSessionId();
+        Long sessionId = service.getSessionId();
         StartProcessCommand startProcessCommand = new StartProcessCommand();
         startProcessCommand.setProcessId( "org.drools.test.TestProcess" );
         ProcessInstance processInstance = service.execute( startProcessCommand );
